--- a/modules/dnn/src/layers/normalize_bbox_layer.cpp
+++ b/modules/dnn/src/layers/normalize_bbox_layer.cpp
@@ -65,7 +65,7 @@ public:
     {
         if (backendId == DNN_BACKEND_INFERENCE_ENGINE)
         {
-            if (!(pnorm == 2 || pnorm == 1)) {
+            if (pnorm != 2) {
                 return false;
             }
             return preferableTarget == DNN_TARGET_MYRIAD ? !acrossSpatial : startAxis == 1;