### This pullrequest changes
The following test now use IE backend:    
Normalize layer
* Test_Torch_layers.net_normalize
* Test_Torch_nets.OpenFace_accuracy

Resize layer
* Test_Darknet_nets.YOLOv3
* Test_Darknet_layers.upsample
* Test_Torch_layers.upsampling_nearest  

LRN layer
* Test_Caffe_layers.LRN

<cut/>

### Performance test results 
Processor: Intel® Core™ i5-6500 CPU @ 3.20GHz × 4
Graphics: Intel® HD Graphics 530 (Skylake GT2)  

| Name of Test                            | Time before changes (ms) | Time after changes (ms) |  x-factor after vs before |
|-----------------------------------------|--------------------------|-------------------------|---------------------------|
| OpenFace::DNNTestNetwork::DLIE/CPU      | 18.298                   | 17.791                  | 0.97                      |
|  OpenFace::DNNTestNetwork::DLIE/OCL     | 8.375                    | 8.051                   | 0.96                      |
| OpenFace::DNNTestNetwork::DLIE/OCL_FP16 | 8.778                    | 7.491                   | 0.85                      |
| YOLOv3::DNNTestNetwork::DLIE/MYRIAD     | 1946.705                 | 1957.809                | 1.01                      |

```
force_builders=Custom
docker_image:Custom=ubuntu-openvino-2018r5:16.04
test_opencl:Custom=ON
buildworker:Custom=linux-2
#test_modules=dnn
allow_multiple_commits=1
```