@dkurt 2019-02-07
It looks like we can enable OCL_FP16 target as well:
https://github.com/opencv/opencv/blob/b068d26fad1de316de4d6357e38afcf2193f6e32/modules/dnn/test/test_torch_importer.cpp#L280-L281

As here:
https://github.com/opencv/opencv/blob/b068d26fad1de316de4d6357e38afcf2193f6e32/modules/dnn/test/test_backends.cpp#L226-L240

P.S. it looks like we need to refactor framework-specific tests and `test_backends.cpp` because there are duplicated runs.