@l-bat 2019-02-15
Yes, I got result with `OMP_WAIT_POLICY=PASSIVE. ` Without this x-factor = 1.04